+++ dino_extinction/__init__.py
@@ -54,10 +54,12 @@ def create_app(env='DEVELOPMENT'):
         from dino_extinction.blueprints import healthcheck
         from dino_extinction.blueprints import battles
         from dino_extinction.blueprints import dinossaurs
+        from dino_extinction.blueprints import robots
 
         app.register_blueprint(healthcheck.bp, url_prefix='/healthcheck')
         app.register_blueprint(battles.bp, url_prefix='/battles')
         app.register_blueprint(dinossaurs.bp, url_prefix='/dinossaurs')
+        app.register_blueprint(robots.bp, url_prefix='/robots')
 
         return app