+++ dino_extinction/blueprints/robots/models.py
@@ -1,7 +1,7 @@
-"""Dinossaurs Models.
+"""Robots Models.
 
 This module contains all the classes and methods regarding our
-dinossaurs blueprint models.
+robots blueprint models.
 
 """
 import pickle
@@ -12,11 +12,11 @@ from dino_extinction.infrastructure import redis
 from . import constants
 
 
-class DinossaurSchema(Schema):
-    """DinosaurSchema Class.
+class RobotSchema(Schema):
+    """RobotSchema Class.
 
     This class is responsible for handling all data regarding our
-    dinossaurs schema.
+    robots schema.
 
     ...
 
@@ -25,11 +25,15 @@ class DinossaurSchema(Schema):
     battle_id : int
         The ID of the battle that you're handling.
 
+    direction : string
+        The direction where your robot is facing (north, south, west or east)
+
     position : tuple
         An X and Y position tuple storing where your dinossaur is.
 
     """
     battle_id = fields.Integer(required=True)
+    direction = fields.String(required=True)
     position = fields.List(fields.Integer, required=True)
 
     @validates('battle_id')
@@ -53,9 +57,30 @@ class DinossaurSchema(Schema):
         if number_of_digits != 4:
             raise ValidationError('The battle ID should be 4 digits long.')
 
+    @validates('direction')
+    def validate_robot_direction(self, data):
+        """Validate the the robot direction.
+
+        This validator checks if the direction that was chosen is any of
+        the following: north, south, east or west.
+
+        ...
+
+        Raises
+        ------
+        ValidationError
+            If the direction is not valid.
+
+        """
+        allowed_directions = ['north', 'south', 'west', 'east']
+
+        if data not in allowed_directions or not data:
+            raise ValidationError(f"The direction must be "
+                                  f"north, south, west or east")
+
     @validates('position')
-    def validate_dinossaur_position(self, data):
-        """Validate the the dinossaur position.
+    def validate_robot_position(self, data):
+        """Validate the the robot position.
 
         This validator checks if the position was provided with both X and Y
         positions declared.
@@ -72,10 +97,10 @@ class DinossaurSchema(Schema):
             raise ValidationError('You must provide xPos and yPos')
 
     @post_load
-    def create_dinossaur(self, data):
-        """Create a new dinossaur.
+    def create_robot(self, data):
+        """Create a new robot.
 
-        This method will create a new dinossaur if you try to dump a new
+        This method will create a new robot if you try to dump a new
         data and all the attributes are valid.
 
         ...
@@ -84,20 +109,22 @@ class DinossaurSchema(Schema):
         ----------
         data : dict
             A dict containing all the data that you are trying to insert into
-            the new battle. Valid keys: battle_id and position.
+            the new battle. Valid keys: battle_id, direction and position.
 
         """
         battle_id = data['battle_id']
+        direction = data['direction']
         position = data['position']
         xPos = position[0] - 1
         yPos = position[1] - 1
 
-        dinossaur_id = self._create_dino_id()
+        robot_id = self._create_robot_id()
 
-        dinossaur = dict()
-        dinossaur['id'] = dinossaur_id
-        dinossaur['type'] = constants.TYPE
-        dinossaur['position'] = position
+        robot = dict()
+        robot['id'] = robot_id
+        robot['type'] = constants.TYPE
+        robot['direction'] = direction
+        robot['position'] = position
 
         raw_battle = redis.instance.get(battle_id)
         if not raw_battle:
@@ -111,17 +138,17 @@ class DinossaurSchema(Schema):
         if board[xPos][yPos]:
             raise ValidationError('This position is not empty')
 
-        battle.setdefault('entities', {}).update({dinossaur_id: dinossaur})
-        board[xPos][yPos] = dinossaur_id
+        battle.setdefault('entities', {}).update({robot_id: robot})
+        board[xPos][yPos] = robot_id
 
         pickled_battle = pickle.dumps(battle)
         redis.instance.set(battle_id, pickled_battle)
 
-        return dinossaur
+        return robot
 
-    def _create_dino_id(self):
+    def _create_robot_id(self):
         r = randint(0000, 9999)
-        return 'D-{0:04d}'.format(r)
+        return 'R-{0:04d}'.format(r)
 
     def _is_not_valid_index(self, x, y, board):
         return x not in range(len(board[0])) or y not in range(len(board))